+++ app/routes/coffees.py
@@ -1,46 +1,86 @@
-from flask import Blueprint, jsonify, request, session
+from flask import Blueprint, jsonify, request, session, Response
 from ..models import Coffee, db
-from ..utils import format_request
+from ..utils import format_request, login_required
 from typing import Sequence
 
 coffees = Blueprint('coffees', __name__)
 
 @coffees.get('/')
-def all_coffees():
-    if session is None or 'user_id' not in session:
-        return {"errors": {"client_error": "Invalid session token"}}, 400
- 
+@login_required
+def all_coffees() -> Response:
+    """
+    Retrieves all coffees that belong to the logged-in user.
+    
+    :return: A response object containing JSONified list of coffee dictionaries
+    :rtype: Response
+    """
     coffees:Sequence[Coffee] = db.session.scalars(db.select(Coffee).filter_by(user_id=session['user_id'])).all()
     return jsonify([coffee.to_dict() for coffee in coffees])
 
+
 @coffees.get('/<int:id>')
-def get_coffee(id):
+def get_coffee(id) -> Response:
+    """
+    Retrieves one coffee, specified by it's id. This can be used by any user
+    
+    :param id: a coffee id
+    :return: A response object containing a JSON coffee dictionary
+    :rtype: Response
+    """
     coffee:Coffee | None = Coffee.query.get(id)
     if not coffee:
-        return {"errors": {"coffee", "Couldn't find the coffee with the requested id"}}, 404
-    return coffee.to_dict()
+        res = jsonify({"errors": {"coffee", "Couldn't find the coffee with the requested id"}})
+        res.status_code = 404
+        return res
+    
+    return jsonify(coffee.to_dict()) # status_code: 200 (auto)
+
 
 @coffees.post('/')
-def save_coffee():
+@login_required
+def save_coffee() -> Response:
+    """
+    Saves a new coffee to the database, automatically associated with the logged-in user
+    
+    :return: A response containing a JSON dictionary of the new coffee
+    :rtype: Response
+    """
     data = format_request(request.json)
-
     coffee = Coffee(**data)
+    coffee.user_id = session['user_id']
+
     db.session.add(coffee)
     db.session.commit()
 
     # Gets the saved instance from the db
     db.session.refresh(coffee)
 
-    return coffee.to_dict()
+    return jsonify(coffee.to_dict()) # status_code: 200 (auto)
+
 
 @coffees.put('/<int:id>')
-def edit_coffee(id):
+@login_required
+def edit_coffee(id) -> Response:
+    """
+    Updates the coffee specified by 'id'. This route expects a request body containing only
+    coffee-relevant keys.
+    
+    :param id: a coffee id
+    :return: A response containing a JSON coffee dictionary representing the updated state
+    :rtype: Response
+    """
     data = format_request(request.json)
 
     coffee:Coffee | None = Coffee.query.get(id)
 
     if not coffee:
-        return {"errors": {"coffee", "Couldn't find the coffee with the requested id"}}, 404
+        res = jsonify({"errors": {"coffee", "Couldn't find the coffee with the requested id"}})
+        res.status_code = 404
+        return res
+    if coffee.user_id != session['user_id']:
+        res = jsonify({"errors": {"client_error": "Users cannot edit coffees that they did not create"}})
+        res.status_code = 403
+        return res
 
     for key in data:
         setattr(coffee, key, data[key])
@@ -48,13 +88,32 @@ def edit_coffee(id):
     db.session.commit()
     db.session.refresh(coffee)
 
-    return coffee.to_dict()
+    return jsonify(coffee.to_dict()) # status_code: 200 (auto)
+
 
 @coffees.delete('/<int:id>')
-def delete_coffee(id):
+@login_required
+def delete_coffee(id) -> Response:
+    """
+    Deletes the coffee with the specified 'id' from the database
+    
+    :param id: a coffee id
+    :return: An empty response with a "204 - successfully deleted" status code
+    :rtype: Response
+    """
     coffee = Coffee.query.get(id)
+    if coffee is None:
+        res = jsonify({"errors": {"client_error": "Couldn't find the selected coffee"}})
+        res.status_code = 400
+        return res
+
+    if coffee.user_id != session['user_id']:
+        res = jsonify({"errors": {"client_error": "Users cannot delete coffees that they did not create"}})
+        res.status_code = 403
+        return res
 
     db.session.delete(coffee)
     db.session.commit()
 
-    return "", 204
+    res = Response(status=204)
+    return res
